Move bulk of CRT simulation to CRTDetSimAlg class
Split CRT simulation into two steps: FillTaggers, and then CreateData.
