+++ sbndcode/CRT/CRTDetSim.h
@@ -14,6 +14,8 @@
 
 #include "lardataalg/DetectorInfo/ElecClock.h"
 
+#include "sbndcode/CRT/CRTSimulation/CRTDetSimAlg.h"
+
 #include <string>
 
 namespace sbnd {
@@ -33,49 +35,9 @@ public:
   std::string fG4ModuleLabel;
 
 private:
-  /**
-   * Get the channel trigger time relative to the start of the MC event.
-   *
-   * @param engine The random number generator engine
-   * @param clock The clock to count ticks on
-   * @param t0 The starting time (which delay is added to)
-   * @param npe Number of observed photoelectrons
-   * @param r Distance between the energy deposit and strip readout end [mm]
-   * @return Trigger clock ticks at this true hit time
-   */
-  uint32_t getChannelTriggerTicks(CLHEP::HepRandomEngine* engine,
-                                /*detinfo::ElecClock& clock,*/
-                                float t0, float npeMean, float r);
 
-  double fGlobalT0Offset;  //!< Time delay fit: Gaussian normalization
-  double fTDelayNorm;  //!< Time delay fit: Gaussian normalization
-  double fTDelayShift;  //!< Time delay fit: Gaussian x shift
-  double fTDelaySigma;  //!< Time delay fit: Gaussian width
-  double fTDelayOffset;  //!< Time delay fit: Gaussian baseline offset
-  double fTDelayRMSGausNorm;  //!< Time delay RMS fit: Gaussian normalization
-  double fTDelayRMSGausShift;  //!< Time delay RMS fit: Gaussian x shift
-  double fTDelayRMSGausSigma;  //!< Time delay RMS fit: Gaussian width
-  double fTDelayRMSExpNorm;  //!< Time delay RMS fit: Exponential normalization
-  double fTDelayRMSExpShift;  //!< Time delay RMS fit: Exponential x shift
-  double fTDelayRMSExpScale;  //!< Time delay RMS fit: Exponential scale
-  double fClockSpeedCRT; //!< Clock speed for the CRT system [MHz]
-  double fNpeScaleNorm;  //!< Npe vs. distance: 1/r^2 scale
-  double fNpeScaleShift;  //!< Npe vs. distance: 1/r^2 x shift
-  double fQ0;  //!< Average energy deposited for mips, for charge scaling [GeV]
-  double fQPed;  //!< ADC offset for the single-peak peak mean [ADC]
-  double fQSlope;  //!< Slope in mean ADC / Npe [ADC]
-  double fQRMS;  //!< ADC single-pe spectrum width [ADC]
-  double fQThreshold;  //!< ADC charge threshold [ADC]
-  double fTResInterpolator;  //!< Interpolator time resolution [ns]
-  double fPropDelay;  //!< Delay in pulse arrival time [ns/m]
-  double fPropDelayError;  //!< Delay in pulse arrival time, uncertainty [ns/m]
-  double fStripCoincidenceWindow;  //!< Time window for two-fiber coincidence [ns]
-  double fTaggerPlaneCoincidenceWindow;  //!< Time window for two-plane coincidence [ticks]
-  double fAbsLenEff;  //!< Effective abs. length for transverse Npe scaling [cm]
-  bool fUseEdep;  //!< Use the true G4 energy deposited, assume mip if false.
-  double fSipmTimeResponse; //!< Minimum time to resolve separate energy deposits [ns]
-  uint32_t fAdcSaturation; //!< Saturation limit per SiPM in ADC counts
   CLHEP::HepRandomEngine& fEngine; //!< Reference to art-managed random-number engine
+  CRTDetSimAlg fDetAlg; //!< Instance of the CRT detector simulation algorithm
 };
 
 }  // namespace crt
