+++ sbndcode/CRT/crtsimmodules_sbnd.fcl
@@ -1,20 +1,17 @@
 BEGIN_PROLOG
 
-sbnd_crtsim: {
-  module_type: "sbndcode/CRT/CRTDetSim"
-  G4ModuleLabel: "genericcrt"
-
-  // Global timing offset [ns]
-  // For Corsika cosmics, use 2.76e6 (see corsika_sbnd.fcl)
+standard_sbnd_crtsimparams: {
+  # Global timing offset [ns]
+  # For Corsika cosmics, use 2.76e6 (see corsika_sbnd.fcl)
   GlobalT0Offset: 0.0
 
-  // Shape parameters for mean time delay vs. Npe
+  # Shape parameters for mean time delay vs. Npe
   TDelayNorm: 4125.74
   TDelayShift: -300.31
   TDelaySigma: 90.392
   TDelayOffset: -1.525
 
-  // Shape parameters for time delay RMS vs. Npe  
+  # Shape parameters for time delay RMS vs. Npe
   TDelayRMSGausNorm: 2.09138
   TDelayRMSGausShift: 7.23993
   TDelayRMSGausSigma: 170.027
@@ -22,44 +19,50 @@ sbnd_crtsim: {
   TDelayRMSExpShift: 75.6183
   TDelayRMSExpScale: 79.3543
 
-  // Clock speed of the CRT system [MHz] (1 ns = 1 tick @ 1000 MHz)
-  // (was 16 MHz for files before v08_25_00)
+  # Clock speed of the CRT system [MHz] (1 ns = 1 tick @ 1000 MHz)
+  # (was 16 MHz for files before v08_25_00)
   ClockSpeedCRT: 1000.
 
-  // Propagation delay [ns/cm]
+  # Propagation delay [ns/cm]
   PropDelay: 0.0061
   fPropDelayError: 0.007
 
-  // Interpolator time resolution [ns]
+  # Interpolator time resolution [ns]
   TResInterpolator: 1.268
 
-  // Model parameters for Npe vs. distance along strip
+  # Model parameters for Npe vs. distance along strip
   NpeScaleNorm: 8.023e7 #5.261e7
   NpeScaleShift: -1085.0
 
-  // Mean deposited charge for mips [GeV]
+  # Mean deposited charge for mips [GeV]
   UseEdep: true
   Q0: 1.75e-3
 
-  // ADC model: Pedestal offset [ADC], slope [ADC/photon], RMS [ADC]
+  # ADC model: Pedestal offset [ADC], slope [ADC/photon], RMS [ADC]
   QPed: 0.0 #63.6
   QSlope: 40.0 #131.9
   QRMS: 80.0 #15.0
   QThreshold: 60.0 #100
 
-  // Coincidence window for two fibers on a strip [ticks = ns]
+  # Coincidence window for two fibers on a strip [ticks = ns]
   StripCoincidenceWindow: 100.0 #30.0
 
-  // Coincidence for hits in a tagger [ticks = ns]
+  # Coincidence for hits in a tagger [ticks = ns]
   TaggerPlaneCoincidenceWindow: 150.0 #5
 
-  // Effective absorption length (for transverse response) [cm]
+  # Effective absorption length (for transverse response) [cm]
   AbsLenEff: 8.5
 
-  // Minimum time between energy deposits that SiPMs can resolve [ns]
+  # Minimum time between energy deposits that SiPMs can resolve [ns]
   SipmTimeResponse: 2.0
 
   AdcSaturation: 4095
+}
+
+sbnd_crtsim: {
+  module_type: "sbndcode/CRT/CRTDetSim"
+  G4ModuleLabel: "genericcrt"
+  DetSimParams: @local::standard_sbnd_crtsimparams
 }
 
 END_PROLOG
